1. Add Global namespace to Global .cpp/.h sources 2. All exception specifications are commented (see HACKING file) 3. Fix several comparison warnings 4. Cosmetics: every source line should be < 80 chars
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/trunk@598 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/FermaNext/TrussUnit/TrussLoad.h
@@ -167,9 +167,9 @@ public:
         return loads;
     }
 
-    virtual size_t countLoads () const
+    virtual int countLoads () const
     {
-        size_t loadsNum = 0;
+        int loadsNum = 0;
         TrussLoadMapConstIter iter = loads.begin();
         for ( ; iter != loads.end(); ++iter ) {
             if ( iter.value()->isEnabled() )
